requirements.txt
vobject
python-dateutil

django-dept-directory
django-webcal
django-latex
